Re-wrote main MySqlTransport to always be in leased mode and killed off the leased transport
