--- Rebus.MySql.Tests/Transport/Contract/MySqlLeaseTransportBasicSendReceive.cs
@@ -1,9 +1,0 @@
-using NUnit.Framework;
-using Rebus.MySql.Tests.Transport.Contract.Factories;
-using Rebus.Tests.Contracts.Transports;
-
-namespace Rebus.MySql.Tests.Transport.Contract
-{
-    [TestFixture, Category(Categories.MySql)]
-    public class MySqlLeaseTransportBasicSendReceive : BasicSendReceive<MySqlLeaseTransportFactory> { }
-}
