+++ Rebus.MySql/MySql/Transport/MySqlTransport.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -7,7 +8,6 @@ using System.Threading.Tasks;
 using MySqlConnector;
 using Rebus.Bus;
 using Rebus.Config;
-using Rebus.Exceptions;
 using Rebus.Extensions;
 using Rebus.Logging;
 using Rebus.Messages;
@@ -36,35 +36,59 @@ namespace Rebus.MySql.Transport
         public const string MessagePriorityHeaderKey = "rbs2-msg-priority";
 
         /// <summary>
+        /// Key for storing the outbound message buffer when performing <seealso cref="Send"/>
+        /// </summary>
+        public const string OutboundMessageBufferKey = "sql-server-transport-leased-outbound-message-buffer";
+
+        /// <summary>
+        /// Size of the leased_by column
+        /// </summary>
+        public const int LeasedByColumnSize = 200;
+
+        /// <summary>
         /// Default delay between executing the background cleanup task
         /// </summary>
         public static readonly TimeSpan DefaultExpiredMessagesCleanupInterval = TimeSpan.FromSeconds(20);
 
         /// <summary>
-        /// Connection provider for obtaining a database connection
+        /// If not specified the default time messages are leased for
         /// </summary>
-        protected readonly IDbConnectionProvider _connectionProvider;
-
-        private readonly IRebusTime _rebusTime;
+        public static readonly TimeSpan DefaultLeaseTime = TimeSpan.FromMinutes(5);
 
         /// <summary>
-        /// Name of the table this transport is using for storage
+        /// If not specified the amount of tolerance workers will allow a message which has already been leased
         /// </summary>
-        protected readonly TableName _receiveTableName;
+        public static readonly TimeSpan DefaultLeaseTolerance = TimeSpan.FromSeconds(30);
 
-        /// <summary>
-        /// Logger
-        /// </summary>
-        protected readonly ILog _log;
-
+        readonly IDbConnectionProvider _connectionProvider;
+        readonly IRebusTime _rebusTime;
+        readonly TableName _receiveTableName;
+        readonly ILog _log;
         readonly IAsyncTask _expiredMessagesCleanupTask;
         readonly bool _autoDeleteQueue;
         bool _disposed;
+        readonly TimeSpan _leaseInterval;
+        readonly TimeSpan _leaseTolerance;
+        readonly bool _automaticLeaseRenewal;
+        readonly TimeSpan _automaticLeaseRenewalInterval;
+        readonly Func<string> _leasedByFactory;
 
         /// <summary>
-        /// Constructs the transport with the given <see cref="IDbConnectionProvider"/>
+        /// Constructor
         /// </summary>
-        public MySqlTransport(IDbConnectionProvider connectionProvider, string inputQueueName, IRebusLoggerFactory rebusLoggerFactory, IAsyncTaskFactory asyncTaskFactory, IRebusTime rebusTime, MySqlTransportOptions options)
+        /// <param name="connectionProvider">A <see cref="IDbConnection"/> to obtain a database connection</param>
+        /// <param name="inputQueueName">Name of the queue this transport is servicing</param>
+        /// <param name="rebusLoggerFactory">A <seealso cref="IRebusLoggerFactory"/> for building loggers</param>
+        /// <param name="asyncTaskFactory">A <seealso cref="IAsyncTaskFactory"/> for creating periodic tasks</param>
+        /// <param name="rebusTime">A <seealso cref="IRebusTime"/> to provide the current time</param>
+        /// <param name="options">Additional options</param>
+        public MySqlTransport(
+            IDbConnectionProvider connectionProvider,
+            string inputQueueName,
+            IRebusLoggerFactory rebusLoggerFactory,
+            IAsyncTaskFactory asyncTaskFactory,
+            IRebusTime rebusTime,
+            MySqlTransportOptions options)
         {
             if (rebusLoggerFactory == null) throw new ArgumentNullException(nameof(rebusLoggerFactory));
             if (asyncTaskFactory == null) throw new ArgumentNullException(nameof(asyncTaskFactory));
@@ -80,6 +104,21 @@ namespace Rebus.MySql.Transport
 
             _expiredMessagesCleanupTask = asyncTaskFactory.Create("ExpiredMessagesCleanup", PerformExpiredMessagesCleanupCycle, intervalSeconds: intervalSeconds);
             _autoDeleteQueue = options.AutoDeleteQueue;
+            _leasedByFactory = options.LeasedByFactory ?? (() => Environment.MachineName);
+            _leaseInterval = options.LeaseInterval ?? DefaultLeaseTime;
+            _leaseTolerance = options.LeaseInterval ?? DefaultLeaseTolerance;
+
+            var automaticLeaseRenewalInterval = options.LeaseAutoRenewInterval;
+
+            if (!automaticLeaseRenewalInterval.HasValue)
+            {
+                _automaticLeaseRenewal = false;
+            }
+            else
+            {
+                _automaticLeaseRenewal = true;
+                _automaticLeaseRenewalInterval = automaticLeaseRenewalInterval.Value;
+            }
         }
 
         /// <summary>
@@ -152,7 +191,9 @@ namespace Rebus.MySql.Transport
                             `visible` DATETIME(6) NOT NULL,
                             `headers` LONGBLOB NOT NULL,
                             `body` LONGBLOB NOT NULL,
-                            `processing` TINYINT(1) NULL,
+                            `leased_until` datetime(6) NULL,
+                            `leased_by` varchar({LeasedByColumnSize}) NULL,
+                            `leased_at` datetime(6) NULL,
                             PRIMARY KEY (`id`)
                         );
                         ----
@@ -161,27 +202,16 @@ namespace Rebus.MySql.Transport
                             `visible` ASC,
                             `id` ASC,
                             `expiration` ASC,
-                            `processing` ASC
+                            `leased_until` DESC
                         );
                         ----
                         CREATE INDEX `idx_expiration` ON {table.QualifiedName} (
-                            `expiration`,
-                            `processing`
+                            `expiration`
                         );");
                 }
 
-                AdditionalSchemaModifications(connection, table);
                 connection.Complete();
             }
-        }
-
-        /// <summary>
-        /// Provides an opportunity for derived implementations to also update the schema
-        /// </summary>
-        /// <param name="connection">Connection to the database</param>
-        /// <param name="table">Name of the table to create schema modifications for</param>
-        protected virtual void AdditionalSchemaModifications(IDbConnection connection, TableName table)
-        {
         }
 
         /// <summary>
@@ -215,43 +245,112 @@ namespace Rebus.MySql.Transport
                 _log.Info("Table {tableName} exists - it will be dropped now", table.QualifiedName);
 
                 connection.ExecuteCommands($"DROP TABLE IF EXISTS {table};");
-                AdditionalSchemaModificationsOnDeleteQueue(connection, table);
                 connection.Complete();
             }
         }
 
         /// <summary>
-        /// Provides an opportunity for derived implementations to also update the schema when the queue is deleted automatically
+        /// Sends the given transport message to the specified logical destination address by adding it to the messages table.
         /// </summary>
-        protected void AdditionalSchemaModificationsOnDeleteQueue(IDbConnection connection, TableName table)
+        public Task Send(string destinationAddress, TransportMessage message, ITransactionContext context)
         {
+            // Send the messages via a buffer, so either all messages get sent at the end of the transaction, or none do.
+            var outboundMessageBuffer = GetOutboundMessageBuffer(context);
+
+            outboundMessageBuffer.Enqueue(new AddressedTransportMessage
+            {
+                DestinationAddress = GetDestinationAddressToUse(destinationAddress, message), Message = message
+            });
+
+            return Task.CompletedTask;
         }
 
         /// <summary>
-        /// Sends the given transport message to the specified destination queue address by adding it to the queue's table.
+        /// Gets the outbound message buffer for sending of messages
         /// </summary>
-        public virtual async Task Send(string destinationAddress, TransportMessage message, ITransactionContext context)
+        /// <param name="context">Transaction context containing the message buffer</param>
+        ConcurrentQueue<AddressedTransportMessage> GetOutboundMessageBuffer(ITransactionContext context)
         {
-            var connection = await GetConnection(context).ConfigureAwait(false);
+            return context.GetOrAdd(OutboundMessageBufferKey, () =>
+            {
+                var outgoingMessages = new ConcurrentQueue<AddressedTransportMessage>();
 
-            var destinationAddressToUse = GetDestinationAddressToUse(destinationAddress, message);
-            try
+                async Task SendOutgoingMessages(ITransactionContext _)
+                {
+                    using (var connection = await _connectionProvider.GetConnectionAsync())
+                    {
+                        while (outgoingMessages.IsEmpty == false)
+                        {
+                            if (outgoingMessages.TryDequeue(out var addressed) == false)
+                            {
+                                break;
+                            }
+
+                            await InnerSend(addressed.DestinationAddress, addressed.Message, connection);
+                        }
+
+                        await connection.CompleteAsync();
+                    }
+                }
+
+                context.OnCommitted(SendOutgoingMessages);
+
+                return outgoingMessages;
+            });
+        }
+
+        /// <summary>
+        /// Performs persistence of a message to the underlying table
+        /// </summary>
+        /// <param name="destinationAddress">Address the message will be sent to</param>
+        /// <param name="message">Message to be sent</param>
+        /// <param name="connection">Connection to use for writing to the database</param>
+        async Task InnerSend(string destinationAddress, TransportMessage message, IDbConnection connection)
+        {
+            var sendTable = TableName.Parse(destinationAddress);
+
+            using (var command = connection.CreateCommand())
             {
-                await InnerSend(destinationAddressToUse, message, connection).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                throw new RebusApplicationException(e, $"Unable to send to destination {destinationAddress}");
+                command.CommandText = $@"
+                    INSERT INTO {sendTable.QualifiedName} (
+                        `headers`,
+                        `body`,
+                        `priority`,
+                        `visible`,
+                        `expiration`
+                    ) VALUES (
+                        @headers,
+                        @body,
+                        @priority,
+                        date_add(date_add(now(6), INTERVAL @visible_total_seconds SECOND), INTERVAL @visible_microseconds MICROSECOND),
+                        date_add(date_add(now(6), INTERVAL @ttl_total_seconds SECOND), INTERVAL @ttl_microseconds MICROSECOND)
+                    );";
+                var headers = message.Headers.Clone();
+                var priority = GetMessagePriority(headers);
+                var visible = GetInitialVisibilityDelay(headers);
+                var ttl = GetTtl(headers);
+
+                // must be last because the other functions on the headers might change them
+                var serializedHeaders = HeaderSerializer.Serialize(headers);
+
+                command.Parameters.Add("headers", MySqlDbType.VarBinary, MathUtil.GetNextPowerOfTwo(serializedHeaders.Length)).Value = serializedHeaders;
+                command.Parameters.Add("body", MySqlDbType.VarBinary, MathUtil.GetNextPowerOfTwo(message.Body.Length)).Value = message.Body;
+                command.Parameters.Add("priority", MySqlDbType.Int32).Value = priority;
+                command.Parameters.Add("visible_total_seconds", MySqlDbType.Int32).Value = (int)visible.TotalSeconds;
+                command.Parameters.Add("visible_microseconds", MySqlDbType.Int32).Value = visible.Milliseconds * 1000;
+                command.Parameters.Add("ttl_total_seconds", MySqlDbType.Int32).Value = (int)ttl.TotalSeconds;
+                command.Parameters.Add("ttl_microseconds", MySqlDbType.Int32).Value = ttl.Milliseconds * 1000;
+                await command.ExecuteNonQueryAsync().ConfigureAwait(false);
             }
         }
 
         /// <summary>
-        /// Receives the next message by querying the input queue table for a message with a recipient matching this transport's <see cref="Address"/>
+        /// Handle retrieving a message from the queue, decoding it, and performing any transaction maintenance.
         /// </summary>
         /// <param name="context">Transaction context the receive is operating on</param>
         /// <param name="cancellationToken">Cancellation token for the receive operation</param>
         /// <returns>A <seealso cref="TransportMessage"/> or <c>null</c> if no message can be dequeued</returns>
-        public virtual Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cancellationToken)
+        public Task<TransportMessage> Receive(ITransactionContext context, CancellationToken cancellationToken)
         {
             // NOTE: This function is specifically NOT implemented as async, for performance reasons. Performance
             // testing has shown that it's actually slower to run this operation async than it is to run it without
@@ -272,10 +371,14 @@ namespace Rebus.MySql.Transport
                                 SELECT id,
                                        headers,
                                        body
-                                FROM {tableName}
+                                FROM {tableName} 
                                 WHERE visible < now(6) AND 
-                                      expiration > now(6) AND
-                                      processing = 0 
+                                      expiration > now(6) AND 
+                                      1 = CASE
+					                    WHEN leased_until is null then 1
+					                    WHEN date_add(date_add(leased_until, INTERVAL @lease_tolerance_total_seconds SECOND), INTERVAL @lease_tolerance_microseconds MICROSECOND) < now(6) THEN 1
+					                    ELSE 0
+				                      END 
                                 ORDER BY priority DESC, 
                                          visible ASC, 
                                          id ASC 
@@ -292,14 +395,22 @@ namespace Rebus.MySql.Transport
                             // Mark the message as being processed within the transaction
                             command.CommandText = $@"
                                 UPDATE {tableName} 
-                                SET processing = 1 
+                                SET leased_until = date_add(date_add(now(6), INTERVAL @lease_total_seconds SECOND), INTERVAL @lease_microseconds MICROSECOND),
+                                    leased_at = now(6),
+                                    leased_by = @leased_by
                                 WHERE id = @message_id";
+                            command.Parameters.Add("lease_total_seconds", MySqlDbType.Int32).Value = (int)_leaseInterval.TotalSeconds;
+                            command.Parameters.Add("lease_microseconds", MySqlDbType.Int32).Value = _leaseInterval.Milliseconds * 1000;
+                            command.Parameters.Add("lease_tolerance_total_seconds", MySqlDbType.Int32).Value = (int)_leaseTolerance.TotalSeconds;
+                            command.Parameters.Add("lease_tolerance_microseconds", MySqlDbType.Int32).Value = _leaseTolerance.Milliseconds * 1000;
+                            command.Parameters.Add("leased_by", MySqlDbType.VarChar, LeasedByColumnSize).Value = _leasedByFactory();
                             command.Parameters.Add("message_id", MySqlDbType.Int64).Value = messageId;
                             command.ExecuteNonQuery();
 
                             // Now apply transaction semantics to clear the message later
-                            ApplyTransactionSemantics(context, messageId);
+                            ApplyLeasedTransactionSemantics(context, messageId);
                         }
+
                         connection.Complete();
                         return Task.FromResult(transportMessage);
                     }
@@ -315,7 +426,7 @@ namespace Rebus.MySql.Transport
         /// Maps a <seealso cref="MySqlDataReader"/> that's read a result from the message table into a <seealso cref="TransportMessage"/>
         /// </summary>
         /// <returns>A <seealso cref="TransportMessage"/> representing the row or <c>null</c> if no row was available</returns>
-        protected static TransportMessage ExtractTransportMessageFromReader(MySqlDataReader reader)
+        static TransportMessage ExtractTransportMessageFromReader(MySqlDataReader reader)
         {
             if (reader.Read() == false)
             {
@@ -332,25 +443,32 @@ namespace Rebus.MySql.Transport
         /// </summary>
         /// <param name="context">Transaction context of the message processing</param>
         /// <param name="messageId">Identifier of the message currently being processed</param>
-        private void ApplyTransactionSemantics(ITransactionContext context, long messageId)
+        void ApplyLeasedTransactionSemantics(ITransactionContext context, long messageId)
         {
+            AutomaticLeaseRenewer renewal = null;
+            if (_automaticLeaseRenewal)
+            {
+                renewal = new AutomaticLeaseRenewer(this, _receiveTableName.QualifiedName, messageId, _connectionProvider, _automaticLeaseRenewalInterval, _leaseInterval);
+            }
+
             context.OnAborted(
                 ctx =>
                 {
+                    renewal?.Dispose();
                     try
                     {
-                        ClearProcessing(messageId);
+                        UpdateLease(_connectionProvider, _receiveTableName.QualifiedName, messageId, null);
                     }
                     catch (Exception ex)
                     {
                         _log.Error(ex, "While Resetting Lease");
                     }
-                }
-            );
+                });
 
             context.OnCommitted(
                 ctx =>
                 {
+                    renewal?.Dispose();
                     try
                     {
                         DeleteMessage(messageId);
@@ -361,17 +479,19 @@ namespace Rebus.MySql.Transport
                     }
 
                     return Task.CompletedTask;
-                }
-            );
+                });
         }
 
         /// <summary>
-        /// Responsible for clearing the processing flag on a message on transaction abort
+        /// Updates a lease with a new leased_until value
         /// </summary>
-        /// <param name="messageId">Identifier of the message currently being processed</param>
-        private void ClearProcessing(long messageId)
+        /// <param name="connectionProvider">Provider for obtaining a connection</param>
+        /// <param name="tableName">Name of the table the messages are stored in</param>
+        /// <param name="messageId">Identifier of the message whose lease is being updated</param>
+        /// <param name="leaseInterval">New lease interval. If <c>null</c> the lease will be released</param>
+        void UpdateLease(IDbConnectionProvider connectionProvider, string tableName, long messageId, TimeSpan? leaseInterval)
         {
-            using (var connection = _connectionProvider.GetConnection())
+            using (var connection = connectionProvider.GetConnection())
             {
                 while (true)
                 {
@@ -379,7 +499,26 @@ namespace Rebus.MySql.Transport
                     {
                         using (var command = connection.CreateCommand())
                         {
-                            command.CommandText = $@"UPDATE {_receiveTableName.QualifiedName} SET processing = 0 WHERE id = {messageId}";
+                            if (leaseInterval.HasValue)
+                            {
+                                command.CommandText = $@"
+                                    UPDATE {tableName}
+                                    SET leased_until = date_add(date_add(now(6), INTERVAL @lease_total_seconds SECOND), INTERVAL @lease_microseconds MICROSECOND)
+                                    WHERE id = @id";
+                                command.Parameters.Add("id", MySqlDbType.Int64).Value = messageId;
+                                command.Parameters.Add("lease_total_seconds", MySqlDbType.Int32).Value = (int)leaseInterval.Value.TotalSeconds;
+                                command.Parameters.Add("lease_microseconds", MySqlDbType.Int32).Value = leaseInterval.Value.Milliseconds * 1000;
+                            }
+                            else
+                            {
+                                command.CommandText = $@"
+                                    UPDATE {tableName}
+                                    SET leased_until = null,
+                                        leased_by = null,
+                                        leased_at = null
+                                    WHERE id = @id";
+                                command.Parameters.Add("id", MySqlDbType.Int64).Value = messageId;
+                            }
                             command.ExecuteNonQuery();
                         }
                         connection.Complete();
@@ -387,7 +526,7 @@ namespace Rebus.MySql.Transport
                     }
                     catch (MySqlException exception) when (exception.ErrorCode == MySqlErrorCode.LockDeadlock)
                     {
-                        // Keep trying if we get a deadlock until it succeeds
+                        // If we get a transaction deadlock here, keep trying until we succeed
                     }
                 }
             }
@@ -397,7 +536,7 @@ namespace Rebus.MySql.Transport
         /// Responsible for deleting the message on transaction commit
         /// </summary>
         /// <param name="messageId">Identifier of the message currently being processed</param>
-        protected void DeleteMessage(long messageId)
+        void DeleteMessage(long messageId)
         {
             using (var connection = _connectionProvider.GetConnection())
             {
@@ -422,9 +561,51 @@ namespace Rebus.MySql.Transport
         }
 
         /// <summary>
+        /// Handles automatically renewing a lease for a given message
+        /// </summary>
+        class AutomaticLeaseRenewer : IDisposable
+        {
+            readonly MySqlTransport _transport;
+            readonly string _tableName;
+            readonly long _messageId;
+            readonly IDbConnectionProvider _connectionProvider;
+            readonly TimeSpan _leaseInterval;
+            Timer _renewTimer;
+
+            public AutomaticLeaseRenewer(MySqlTransport transport, string tableName, long messageId, IDbConnectionProvider connectionProvider, TimeSpan renewInterval, TimeSpan leaseInterval)
+            {
+                _transport = transport;
+                _tableName = tableName;
+                _messageId = messageId;
+                _connectionProvider = connectionProvider;
+                _leaseInterval = leaseInterval;
+                _renewTimer = new Timer(RenewLease, null, renewInterval, renewInterval);
+            }
+
+            public void Dispose()
+            {
+                _renewTimer?.Change(TimeSpan.FromMilliseconds(-1), TimeSpan.FromMilliseconds(-1));
+                _renewTimer?.Dispose();
+                _renewTimer = null;
+            }
+
+            void RenewLease(object state)
+            {
+                try
+                {
+                    _transport.UpdateLease(_connectionProvider, _tableName, _messageId, _leaseInterval);
+                }
+                catch (Exception ex)
+                {
+                    _transport._log.Error(ex, "While Renewing Lease");
+                }
+            }
+        }
+
+        /// <summary>
         /// Gets the address a message will actually be sent to. Handles deferred messages.
         /// </summary>
-        protected static string GetDestinationAddressToUse(string destinationAddress, TransportMessage message)
+        static string GetDestinationAddressToUse(string destinationAddress, TransportMessage message)
         {
             return string.Equals(destinationAddress, MagicExternalTimeoutManagerAddress, StringComparison.CurrentCultureIgnoreCase)
                 ? GetDeferredRecipient(message)
@@ -439,53 +620,6 @@ namespace Rebus.MySql.Transport
             }
 
             throw new InvalidOperationException($"Attempted to defer message, but no '{Headers.DeferredRecipient}' header was on the message");
-        }
-
-        /// <summary>
-        /// Performs persistence of a message to the underlying table
-        /// </summary>
-        /// <param name="destinationAddress">Address the message will be sent to</param>
-        /// <param name="message">Message to be sent</param>
-        /// <param name="connection">Connection to use for writing to the database</param>
-        protected async Task InnerSend(string destinationAddress, TransportMessage message, IDbConnection connection)
-        {
-            var sendTable = TableName.Parse(destinationAddress);
-
-            using (var command = connection.CreateCommand())
-            {
-                command.CommandText = $@"
-                    INSERT INTO {sendTable.QualifiedName} (
-                        `headers`,
-                        `body`,
-                        `priority`,
-                        `visible`,
-                        `expiration`,
-                        `processing`
-                    ) VALUES (
-                        @headers,
-                        @body,
-                        @priority,
-                        date_add(date_add(now(6), INTERVAL @visible_total_seconds SECOND), INTERVAL @visible_microseconds MICROSECOND),
-                        date_add(date_add(now(6), INTERVAL @ttl_total_seconds SECOND), INTERVAL @ttl_microseconds MICROSECOND),
-                        0
-                    );";
-                var headers = message.Headers.Clone();
-                var priority = GetMessagePriority(headers);
-                var visible = GetInitialVisibilityDelay(headers);
-                var ttl = GetTtl(headers);
-
-                // must be last because the other functions on the headers might change them
-                var serializedHeaders = HeaderSerializer.Serialize(headers);
-
-                command.Parameters.Add("headers", MySqlDbType.VarBinary, MathUtil.GetNextPowerOfTwo(serializedHeaders.Length)).Value = serializedHeaders;
-                command.Parameters.Add("body", MySqlDbType.VarBinary, MathUtil.GetNextPowerOfTwo(message.Body.Length)).Value = message.Body;
-                command.Parameters.Add("priority", MySqlDbType.Int32).Value = priority;
-                command.Parameters.Add("visible_total_seconds", MySqlDbType.Int32).Value = (int)visible.TotalSeconds;
-                command.Parameters.Add("visible_microseconds", MySqlDbType.Int32).Value = visible.Milliseconds * 1000;
-                command.Parameters.Add("ttl_total_seconds", MySqlDbType.Int32).Value = (int)ttl.TotalSeconds;
-                command.Parameters.Add("ttl_microseconds", MySqlDbType.Int32).Value = ttl.Milliseconds * 1000;
-                await command.ExecuteNonQueryAsync().ConfigureAwait(false);
-            }
         }
 
         TimeSpan GetInitialVisibilityDelay(IDictionary<string, string> headers)
@@ -542,8 +676,7 @@ namespace Rebus.MySql.Transport
                         command.CommandText = $@"
                             SELECT id
                             FROM {_receiveTableName.QualifiedName}
-                            WHERE expiration < now() and
-                                  processing = 0
+                            WHERE expiration < now()
                             LIMIT 100";
                         using (var reader = await command.ExecuteReaderAsync().ConfigureAwait(false))
                         {
@@ -573,16 +706,6 @@ namespace Rebus.MySql.Transport
                 _log.Info("Performed expired messages cleanup in {cleanupTimeSeconds} - {expiredMessageCount} expired messages with recipient {queueName} were deleted",
                     stopwatch.Elapsed.TotalSeconds, results, _receiveTableName.QualifiedName);
             }
-        }
-
-        async Task<IDbConnection> GetConnection(ITransactionContext context)
-        {
-            // Get the connection and set it up to be disposed or committed in the transaction context. MySQL cannot
-            // share connections, so we do not store it in the shared context items, but create a new one each time.
-            var connection = await _connectionProvider.GetConnectionAsync().ConfigureAwait(false);
-            context.OnCommitted(async ctx => await connection.CompleteAsync().ConfigureAwait(false));
-            context.OnDisposed(ctx => connection.Dispose());
-            return connection;
         }
 
         static int GetMessagePriority(Dictionary<string, string> headers)
@@ -617,6 +740,12 @@ namespace Rebus.MySql.Transport
             {
                 _disposed = true;
             }
+        }
+
+        class AddressedTransportMessage
+        {
+            public string DestinationAddress { get; set; }
+            public TransportMessage Message { get; set; }
         }
     }
 }
